\name{tCFA-package}
\alias{tCFA-package}
\alias{tCFA}
\docType{package}
\title{
\packageTitle{tCFA}
}
\description{
\packageDescription{tCFA}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{tCFA}
\packageIndices{tCFA}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{tCFA}

Maintainer: \packageMaintainer{tCFA}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
